backend/requirements.txt
dj-rest-auth==5.0.2
Django==5.0.2
django-allauth==0.57.2
django-extensions==3.2.3
django-cachalot==2.6.2
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-filter==23.5
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-storages[s3]==1.14.2
drf-yasg==1.21.7
gunicorn==21.2.0
psycopg2==2.9.9
redis==5.0.1
rest-framework-roles==1.0.3
